[controllers] Refactor: Moving controllers within pkg directory
